PROJ_40 refactor: rename folder for set password
