+++ features/newPassword/ui/NewPasswordForm.module.scss
